+++ CHANGELOG.md
@@ -12,7 +12,7 @@
 
 ### Upgrade guide
 1. Deploy new image
-    1. No database migration required
+2. Migrate database